backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
email-validator>=2.0.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
psycopg[binary]>=3.1.18
requests>=2.31.0
httpx>=0.24.0
pystac-client>=0.7.0
rasterio>=1.3.0
numpy>=1.24.0
boto3>=1.28.0
jinja2>=3.1.0
google-generativeai>=0.3.0
reportlab>=4.0.0
